feat: Add multiple docker-compose setup
Each platform/tool now has it's own docker-compose.yml file. We stich
all the docker-compose files together using a strategy from this gist:
https://gist.github.com/Gems/a40d7eb45f46c82f990aa7e8845d7e7a. We can
write a Makefile/set of scripts/tool in the future that will
automatically stitch together the correct docker-compose setups when we
are measuring. Still need to solve how we will pass in the different env
variables.