feat: Automatically run docker-compose
I spent a while with different approaches to try to get compose-go
or compose-cli working, but I couldn't get it going without having to go
through the trouble of setting up a custom client and backend. For now,
I'm just running docker-compose through exec.Cmd with a temporary
docker-compose file that I built.