Rewrite tool/loadgen
The rewrite aims to make it easier to use the loadgen outside of the
containerized environment setup by sentry-sdk-benchmark, thus making it easier
to debug.

Now loadgen uses command-line flags instead of environment variables,
and exposes a few more knobs, replacing hard-coded values (e.g. URLs).

Logging has been revisited to have more consistent messages covering all
operations and including relevant information.

In case of any error, the file name and line number of the error is now
included in the output.

Print a text report of the test run to stdout.

Write all output to a dedicated directory, instead of interpreting path
as "path/to/dir/prefix" and creating files under "path/to/dir" that are
named with the prefix "prefix".