Make it possible to run a single scenario
When trying out changes or adding a new platform, it is useful to be
able to run only the baseline or instrumented scenario instead of being
forced to run both, or resort to editing code to disable one of the
other.

Report generation was also changed such that now we can generate a
report for a single scenario.