Add Django app instrumented with OpenTelemetry
This is to be able to compare the Sentry instrumented app with something else
to better make sense of the difference between baseline and instrumented.

We choose to use the Zipkin JSON exporter because that is closest to what the
Sentry SDK does. The Sentry SDK sends a JSON payload over HTTP.

We also include database instrumentation, since Sentry offers that
out-of-the-box.