bench: expose maxwait config for Java/Spring
Often times the Spring app takes a long time to boot, going over the
default value for maxwait.

Instead of bumping maxwait for all (a low maxwait means we can fail
faster if an app has problems booting up), we expose the config and bump
the value only for the Spring app.