bench: expose flag to prevent opening a browser
An open browser can interfere with benchmark runs. Sometimes it is
convenient or necessary to run multiple benchmarks and avoid opening a
browser in the middle of the runs.

Sometimes the tool is run in an environment where no browser is
available or relevant, for example when running on a remote host.