dygraph: remove logscale option
A typo on the option name in JSON prevents the option from being applied
(correct is "logscale" all lower case).

But we, in fact, don't want the Latency Over Time graph to have logscale
axes.

Considering that we don't try to have a full Dygraph wrapper in Go, the
simpler path forward is to remove the option, instead of fixing the typo
and setting the option to false.